final commit to mandi?

+++ COHORT/v_COHORT$identifiers_long.sql
@@ -6,7 +6,7 @@ ALTER VIEW COHORT$identifiers_long AS
 SELECT co.schoolid
       ,CASE
         WHEN sp.program_name = 'Out of District' THEN sp.programid
-        ELSE CONVERT(INT,CONCAT(co.schoolid, sp.programid)) 
+        ELSE CONVERT(BIGINT,CONCAT(co.schoolid, sp.programid)) 
        END AS reporting_schoolid
       ,CASE
         WHEN co.schoolid = 73252 THEN 'MS'
